wip: adds new fields, thinking about combination output (odometer, preparer)
