+++ src/app/forms/components/dynamic-form-group/dynamic-form-group.component.ts
@@ -1,6 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { FormGroup } from '@angular/forms';
-import { CustomFormArray, CustomFormGroup, DynamicFormService, FormNode, FormNodeTypes } from '../../services/dynamic-form.service';
+import { DynamicFormService } from '../../services/dynamic-form.service';
+import { CustomFormArray, CustomFormGroup, FormNode, FormNodeTypes } from '../../services/dynamic-form.types';
 
 @Component({
   selector: 'app-dynamic-form-group',
