+++ src/app/forms/services/dynamic-form.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { DynamicFormService, FormNode, FormNodeTypes, CustomFormControl, FormNodeViewTypes, CustomControl, CustomFormGroup } from './dynamic-form.service';
-import { AbstractControl, FormArray, FormControl, FormGroup, ValidatorFn, Validators } from '@angular/forms';
+import { AbstractControl, FormArray, ValidatorFn, Validators } from '@angular/forms';
 
 describe('DynamicFormService', () => {
   let service: DynamicFormService;
@@ -36,7 +36,8 @@ describe('DynamicFormService', () => {
             name: 'vin',
             label: 'Vechile Identification Number',
             type: FormNodeTypes.CONTROL,
-            viewType: FormNodeViewTypes.STRING
+            viewType: FormNodeViewTypes.STRING,
+            children: []
           }
         ]
       };
@@ -65,7 +66,8 @@ describe('DynamicFormService', () => {
                 name: 'vin',
                 label: 'Vechile Identification Number',
                 type: FormNodeTypes.CONTROL,
-                viewType: FormNodeViewTypes.STRING
+                viewType: FormNodeViewTypes.STRING,
+                children: []
               }
             ]
           }
