+++ src/app/models/country-of-registration.enum.ts
@@ -1,0 +1,40 @@
+export enum CountryOfRegistration {
+  GreatBritainandNorthernIreland_GB = 'gb'
+  // 'Alderney - GBA' = 'gba',
+  // 'Guernsey - GBG' = 'gbg',
+  // 'Jersey - GBJ' = 'gbj',
+  // 'Isle of Man - GBM' = 'gbm',
+  // 'Gibraltar - GBZ' = 'gbz',
+  // 'Austria - A' = 'a',
+  // 'Belgium - B' = 'b',
+  // 'Bosnia and Herzegovina - BIH' = 'bih',
+  // 'Bulgaria - BG' = 'bg',
+  // 'Croatia - HR' = 'hr',
+  // 'Cyprus - CY' = 'cy',
+  // 'Czech Republic - CZ' = 'cz',
+  // 'Denmark - DK' = 'dk',
+  // 'Estonia - EST' = 'est',
+  // 'Finland - FIN' = 'fin',
+  // 'France - F' = 'f',
+  // 'Germany - D' = 'd',
+  // 'Greece - GR' = 'gr',
+  // 'Hungary - H' = 'h',
+  // 'Ireland - IRL' = 'irl',
+  // 'Italy - I' = 'i',
+  // 'Latvia - LV' = 'lv',
+  // 'Lithuania - LT' = 'lt',
+  // 'Luxembourg - L' = 'l',
+  // 'Malta - M' = 'm',
+  // 'Netherlands - NL' = 'nl',
+  // 'Norway - N' = 'n',
+  // 'Poland - PL' = 'pl',
+  // 'Portugal - P' = 'p',
+  // 'Romania - RO' = 'ro',
+  // 'Slovakia - SK' = 'sk',
+  // 'Slovenia - SLO' = 'slo',
+  // 'Spain - E' = 'e',
+  // 'Sweden - S' = 's',
+  // 'Switzerland - CH' = 'ch',
+  // 'Non EU' = 'non-eu',
+  // 'Country Not Known' = 'not-known'
+}
