+++ src/app/models/eu-vehicle-category.enum.ts
@@ -1,0 +1,20 @@
+export enum EuVehicleCategory {
+  M1 = 'm1',
+  M2 = 'm2',
+  M3 = 'm3',
+  N1 = 'n1',
+  N2 = 'n2',
+  N3 = 'n3',
+  O1 = 'o1',
+  O2 = 'o2',
+  O3 = 'o3',
+  O4 = 'o4',
+  L1E_A = 'l1e-a',
+  L1E = 'l1e',
+  L2E = 'l2e',
+  L3E = 'l3e',
+  L4E = 'l4e',
+  L5E = 'l5e',
+  L6E = 'l6e',
+  L7E = 'l7e'
+}
