+++ src/app/models/odometer-unit.enum.ts
@@ -1,0 +1,4 @@
+export enum OdometerReadingUnits {
+  KILOMETERS = 'kilometers',
+  MILES = 'miles'
+}
